+++ esriUtil/esriUtil/FunctionRasters/NeighborhoodHelper/aggregationHelperMax.cs
@@ -8,7 +8,7 @@ namespace esriUtil.FunctionRasters.NeighborhoodHelper
 {
     class aggregationHelperMax : aggregationFunctionDataset
     {
-        public override object getTransformedValue(System.Array bigArr, int startClms, int startRws, int cells, double noDataValue)
+        public override object getTransformedValue(System.Array bigArr, int startClms, int startRws, int cells, float noDataValue)
         {
             return blockHelperStats.getBlockMax(bigArr, startClms, startRws, cells, noDataValue);
         }
