+++ esriUtil/esriUtil/FunctionRasters/NeighborhoodHelper/focalSampleHelperMedian.cs
@@ -10,12 +10,12 @@ namespace esriUtil.FunctionRasters.NeighborhoodHelper
     {
         public override object getTransformedValue(System.Array bigArr, int startClm, int startRw)
         {
-            List<double> dbLst = new List<double>();
+            List<float> dbLst = new List<float>();
             foreach (int[] xy in offsetLst)
             {
                 int bWc = xy[0] + startClm;
                 int bRc = xy[1] + startRw;
-                double vl = System.Convert.ToDouble(bigArr.GetValue(bWc, bRc));
+                float vl = System.Convert.ToSingle(bigArr.GetValue(bWc, bRc));
                 if (rasterUtil.isNullData(vl, noDataValue))
                 {
                     continue;
