+++ esriUtil/esriUtil/FunctionRasters/NeighborhoodHelper/focalSampleHelperMode.cs
@@ -10,12 +10,12 @@ namespace esriUtil.FunctionRasters.NeighborhoodHelper
     {
         public override object getTransformedValue(System.Array bigArr, int startClm, int startRw)
         {
-            Dictionary<double, int> countDic = new Dictionary<double, int>();
+            Dictionary<float, int> countDic = new Dictionary<float, int>();
             foreach (int[] xy in offsetLst)
             {
                 int bWc = xy[0] + startClm;
                 int bRc = xy[1] + startRw;
-                double vl = System.Convert.ToDouble(bigArr.GetValue(bWc, bRc));
+                float vl = System.Convert.ToSingle(bigArr.GetValue(bWc, bRc));
                 if (rasterUtil.isNullData(vl, noDataValue))
                 {
                     continue;
@@ -36,10 +36,10 @@ namespace esriUtil.FunctionRasters.NeighborhoodHelper
                 }
             }
             int vlMax = countDic.Values.Max();
-            double key = noDataValue;
-            foreach (KeyValuePair<double, int> kVp in countDic)
+            float key = noDataValue;
+            foreach (KeyValuePair<float, int> kVp in countDic)
             {
-                double k = kVp.Key;
+                float k = kVp.Key;
                 int v = kVp.Value;
                 if (v == vlMax)
                 {
