+++ esriUtil/esriUtil/FunctionRasters/NeighborhoodHelper/focalSampleHelperStd.cs
@@ -11,14 +11,14 @@ namespace esriUtil.FunctionRasters.NeighborhoodHelper
         public override object getTransformedValue(System.Array bigArr, int startClm, int startRw)
         {
             //Console.WriteLine("Start CR = " + startClm.ToString()+":"+ startRw.ToString());
-            double s = 0;
-            double s2 = 0;
+            float s = 0;
+            float s2 = 0;
             foreach (int[] xy in offsetLst)
             {
                 int bWc = xy[0] + startClm;
                 int bRc = xy[1] + startRw;
 
-                double vl = System.Convert.ToDouble(bigArr.GetValue(bWc, bRc));
+                float vl = System.Convert.ToSingle(bigArr.GetValue(bWc, bRc));
                 //Console.WriteLine("\t"+vl.ToString());
                 if (rasterUtil.isNullData(vl, noDataValue))
                 {
