+++ esriUtil/esriUtil/FunctionRasters/NeighborhoodHelper/neighborhoodHelperLandscapeMaxAreaRectangle.cs
@@ -8,7 +8,7 @@ namespace esriUtil.FunctionRasters.NeighborhoodHelper
 {
     class neighborhoodHelperLandscapeMaxAreaRectangle: neighborhoodHelperLandscapeRectangleBase
     {
-        public override double findUniqueRegionsValue(Dictionary<int, int[]> uniqueDic)
+        public override float findUniqueRegionsValue(Dictionary<int, int[]> uniqueDic)
         {
             int maxVl = 0;
             foreach (int[] cntArr in uniqueDic.Values)
