+++ esriUtil/esriUtil/FunctionRasters/NeighborhoodHelper/neighborhoodHelperLandscapeMeanEdgeRectangle.cs
@@ -8,13 +8,13 @@ namespace esriUtil.FunctionRasters.NeighborhoodHelper
 {
     class neighborhoodHelperLandscapeMeanEdgeRectangle: neighborhoodHelperLandscapeRectangleBase
     {
-        public override double findUniqueRegionsValue(Dictionary<int, int[]> uniqueDic)
+        public override float findUniqueRegionsValue(Dictionary<int, int[]> uniqueDic)
         {
-            double n = uniqueDic.Count;
-            double sum = 0;
+            float n = uniqueDic.Count;
+            float sum = 0;
             foreach (int[] cntArr in uniqueDic.Values)
             {
-                double vl = cntArr[1];
+                float vl = cntArr[1];
                 sum += vl;
             }
             return sum / n;
