+++ esriUtil/esriUtil/FunctionRasters/NeighborhoodHelper/neighborhoodHelperLandscapeModeRatioRectangle.cs
@@ -8,12 +8,12 @@ namespace esriUtil.FunctionRasters.NeighborhoodHelper
 {
     class neighborhoodHelperLandscapeModeRatioRectangle : neighborhoodHelperLandscapeRectangleBase
     {
-        public override double findUniqueRegionsValue(Dictionary<int, int[]> uniqueDic)
+        public override float findUniqueRegionsValue(Dictionary<int, int[]> uniqueDic)
         {
-            Dictionary<double,int> ratioDic = new Dictionary<double,int>();
+            Dictionary<float,int> ratioDic = new Dictionary<float,int>();
             foreach (int[] cntArr in uniqueDic.Values)
             {
-                double ratio = System.Convert.ToDouble(cntArr[1]) / System.Convert.ToDouble(cntArr[0]);
+                float ratio = System.Convert.ToSingle(cntArr[1]) / System.Convert.ToSingle(cntArr[0]);
                 int cnt;
                 if (ratioDic.TryGetValue(ratio, out cnt))
                 {
@@ -24,11 +24,11 @@ namespace esriUtil.FunctionRasters.NeighborhoodHelper
                     ratioDic.Add(ratio, 1);
                 }
             }
-            double mode = 0;
+            float mode = 0;
             int cntMax = ratioDic.Values.Max();
-            foreach (KeyValuePair<double, int> kvp in ratioDic)
+            foreach (KeyValuePair<float, int> kvp in ratioDic)
             {
-                double k = kvp.Key;
+                float k = kvp.Key;
                 int v = kvp.Value;
                 if (v == cntMax)
                 {
